move delegate action into separate file
